debug attack & ability anims in sprite-viewer
