+++ app/public/src/pages/component/debug/debug-scene.css
@@ -1,5 +1,16 @@
 #debug-scene {
   position: relative;
+  width: 800px;
+  height: 800px;
+  overflow: hidden;
+  border: 2px solid black;
+  border-radius: 4px;
+}
+
+#debug-scene canvas {
+  position: relative;
+  left: -580px;
+  top: 0;
 }
 
 #debug-scene #status-message {
@@ -8,6 +19,7 @@
   top: 0;
   color: white;
   background-color: black;
+  z-index: 2;
 }
 
 #debug-scene-controls {
